pyproject.toml
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "venv"]
line-length = 99
indent-width = 4
target-version = "py312"
select = [
  # "ALL"
  "E",   # pycodestyle (`E`)
  "F",   # Pyflakes (`F`)
  "I",   # isort (`I`)
  "ANN", # flake8-annotations (`ANN`)
  "PLW", # Warning (`PLW`)
  "UP",  # pyupgrade (`UP`)
  "A",   # flake8-builtins (`A`)
  "N",   # pep8-naming (`N`)
  "D",   # pydocstyle (`D`)
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
line-ending = "auto"

[tool.poetry]
name = "app"
version = "0.1.0"
description = "test cloud run jobs project"
authors = ["Your Name <you@example.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.25.0"
google-cloud-logging = "^3.11.2"
google-cloud-bigquery-storage = "^2.26.0"
msgpack = "1.0.8"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
